"Cambio de funciones de ayuda y corrección de errores en el menú, creación de archivos .hxx y .cpp para comandos del menú.\n\n" "Agregado el archivo .gitignore para ignorar archivos de salida (Ejecutables) y configuraciones de VSCode.\n\n"
