+++ .gitignore
@@ -1,0 +1,4 @@
+/output
+*.exe
+.vscode
+.vscode/*
