+++ menu_commands.hxx
@@ -1,0 +1,12 @@
+//Declaración para el menú de comandos
+#ifndef MENU_COMMANDS_H
+#define MENU_COMMANDS_H
+
+#include <iostream>
+
+void comandoAyuda(std::string argumento, int momento);
+void comandoNoReconocido();
+
+
+
+#endif
